refactor: make membership service less visible
